pyproject.toml
[project]
name = "dados-tabnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "python-slugify>=8.0.4",
]
